add react-router and protected dashboard redirect
