--- client/src/components/App/index.ts
@@ -1,1 +1,0 @@
-export { default } from './App'
