+++ client/src/components/Lists/Lists.tsx
@@ -19,7 +19,7 @@ class Lists extends Component<Props> {
           Create a new List
         </Button>
         <Pane marginY={scale(4)}>
-          <Heading size={100} marginBlock={scale(1)}>Your Lists</Heading>
+          <Heading size={100} marginBottom={scale(1)}>Your Lists</Heading>
           <Pane is='ul' marginLeft={0} paddingLeft={0}>
             {this.props.lists.map(list => (
               <Pane is='li' listStyle='none' marginY={scale(1)} key={list.id}>
