+++ client/src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import './index.css'
-import App from './components/App'
+import App from './App'
 
 ReactDOM.render(<App />, document.getElementById('root'))
